Moving extract code into a dedicated module
